Seperated components and moved files into a folder
